pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "jwtcli"
authors = [{name = "Alexander O'Donovan-Jones"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
description = "foo"
dynamic = ["version"]
dependencies = [
    "pyjwt==2.8.0",
    "click==8.1.7",
    "click-default-group==1.2.4"
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.1",
    "ruff==0.2.2",
    "flit-core==3.9.0",
    "build==1.0.3",
    "pre-commit==3.6.1",
    "ipython==8.22.1",
    "pyyaml==6.0.1",
    "jinja2==3.1.3"
]

[project.urls]
Home = "https://github.com/aodj/jwt"

[project.scripts]
jwt = "jwtcli.cli:cli"
